Move ErrorResponse into the common lib
